fix: chyby jen na MySQL — emoji, hledání, součty, duplicity
* Reakce emoji: utf8mb4_unicode_ci váží všechna emoji stejně, takže
  😂 po 👍 smazalo 👍. Migrace dá sloupci chat_reactions.emoji
  utf8mb4_bin (na SQLite nic).
* FULLTEXT v booleovském režimu padal na @ a jiných operátorech
  a ignoroval krátká slova. FulltextDotaz operátory odstraní, krátká
  slova a stopslova vynechá, zbytek hledá jako „všechna, i začátkem
  slova"; bez použitelného slova zůstává LIKE.
* SUM() vrací na MySQL řetězec — časová osa, statistiky, rozpočet akce,
  úložiště a obnova posílaly do JSON "3" místo 3.
* Stejný název rozpočtu dárků (i jen jinak velkými písmeny nebo bez
  diakritiky) → 500; teď 422.
* insertOrIgnore bez klíče, který by duplicitu chytil: přílohy
  vzpomínkového večera a připomínky sdíleného úkolu se mohly zdvojit
  (dvě staré kopie úkolu založily dvě akce); potvrzení asistenta
  schovávalo i jiné chyby než duplicitu.

+++ app/Http/Controllers/RecoveryController.php
@@ -192,19 +192,20 @@ class RecoveryController extends Controller
             [
                 'key' => 'screenshots', 'label' => 'Screenshoty', 'icon' => '📱',
                 'count' => (clone $base)->where(fn ($query) => $query->where('original_filename', 'like', '%screenshot%')->orWhere('original_filename', 'like', '%screen shot%'))->count(),
-                'bytes' => (clone $base)->where(fn ($query) => $query->where('original_filename', 'like', '%screenshot%')->orWhere('original_filename', 'like', '%screen shot%'))->sum('size_bytes'),
+                // `(int)`: `sum()` je v MySQL DECIMAL a z PDO přijde řetězec, na SQLite číslo.
+                'bytes' => (int) (clone $base)->where(fn ($query) => $query->where('original_filename', 'like', '%screenshot%')->orWhere('original_filename', 'like', '%screen shot%'))->sum('size_bytes'),
                 'reason' => 'Názvy souborů odpovídají screenshotům; před úklidem je vždy zkontrolujte.', 'action' => '/search?q=screenshot',
             ],
             [
                 'key' => 'large_videos', 'label' => 'Velká videa', 'icon' => '🎬',
                 'count' => (clone $base)->where('media_type', 'video')->where('size_bytes', '>=', 500 * 1024 * 1024)->count(),
-                'bytes' => (clone $base)->where('media_type', 'video')->where('size_bytes', '>=', 500 * 1024 * 1024)->sum('size_bytes'),
+                'bytes' => (int) (clone $base)->where('media_type', 'video')->where('size_bytes', '>=', 500 * 1024 * 1024)->sum('size_bytes'),
                 'reason' => 'Videa větší než 500 MB mají největší dopad na úložiště.', 'action' => '/search?media_type=video&sort_by=size_bytes',
             ],
             [
                 'key' => 'unorganized', 'label' => 'Nezařazené', 'icon' => '📥',
                 'count' => (clone $base)->whereNull('primary_album_id')->whereDoesntHave('albums')->count(),
-                'bytes' => (clone $base)->whereNull('primary_album_id')->whereDoesntHave('albums')->sum('size_bytes'),
+                'bytes' => (int) (clone $base)->whereNull('primary_album_id')->whereDoesntHave('albums')->sum('size_bytes'),
                 'reason' => 'Položky nejsou v žádném albu; můžete je roztřídit nebo archivovat.', 'action' => '/inbox',
             ],
             [
